// lib/api/generated/model/couponDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */
import type { DiscountType } from "./discountType";

export interface CouponDto {
  id?: number;
  /** @nullable */
  code?: string | null;
  discountValue?: number;
  discountType?: DiscountType;
  startDate?: string;
  endDate?: string;
  usageLimit?: number;
  usageCount?: number;
  isActive?: boolean;
  businessId?: number;
}