// lib/api/generated/model/createCouponCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */
import type { DiscountType } from "./discountType";

export interface CreateCouponCommand {
  businessId?: number;
  /** @nullable */
  code?: string | null;
  discountValue?: number;
  discountType?: DiscountType;
  startDate?: string;
  endDate?: string;
  usageLimit?: number;
  isActive?: boolean;
}