// lib/api/generated/model/discountType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export type DiscountType = (typeof DiscountType)[keyof typeof DiscountType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DiscountType = {
  Percentage: "Percentage",
  FixedAmount: "FixedAmount",
} as const;